requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
openai
unstructured
unstructured[local-inference]
pinecone-client==2.2.4
poppler-utils
huggingface-hub
langchain_community
qdrant_client
fastembed
tqdm
